refactor: move entity-related exceptions to registry package
Resolves #40
