+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/registry/DataAccessException.java
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: Apache-2.0 */
-package hu.bme.mit.ftsrg.hypernate.entity;
+package hu.bme.mit.ftsrg.hypernate.registry;
 
 import hu.bme.mit.ftsrg.hypernate.HypernateException;
 import lombok.experimental.StandardException;
