+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/registry/MissingPrimaryKeysException.java
@@ -1,0 +1,8 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+package hu.bme.mit.ftsrg.hypernate.registry;
+
+import hu.bme.mit.ftsrg.hypernate.HypernateException;
+import lombok.experimental.StandardException;
+
+@StandardException
+public class MissingPrimaryKeysException extends HypernateException {}
